requirements.txt
langgraph
langchain
langchain_openai
langchain-google-genai
google-generativeai
pyparsing<3.0.0
pydantic
psycopg2-binary
sqlalchemy
v20
python-dotenv
streamlit
plotly
langchain-google-genai
langchain
langchain_core
langchain_community
pandas
